Mostly more Add Data
This will allow user to add new dive site based on zip code search. It
will create a new dive log on a new dive site, but has a problem
associating the dive log with a user. Does NOT successfully create a new
dive log to existing dive site yet.